+++ lib/screengem.rb
@@ -7,6 +7,7 @@ require "active_support/core_ext"
 require "capybara/dsl"
 require "rspec/expectations"
 
+require "screengem/automatic_visit"
 require "screengem/incorrect_answer"
 require "screengem/screen_element"
 require "screengem/screengem_error"