+++ spec/screengem/screen_elements_spec.rb
@@ -1,14 +1,28 @@
 module Screengem
   RSpec.describe ScreenElements do
     Module.new do
-      ExamplePage = Class.new(Screengem::ScreenElement)
-      ExampleSection = Class.new(Screengem::ScreenElement)
+      # Does not override the visit_path method
+      ExampleOne = Class.new(Screengem::ScreenElement)
+
+      # Does override the visit_path method
+      ExampleTwo = Class.new(Screengem::ScreenElement) do
+        def visit_path
+          "some/path"
+        end
+      end
     end
 
-    it "creates screen element methods" do
-      generated_creation_methods = Screengem::ScreenElements.instance.public_methods(false)
+    subject(:screen) { Screengem::ScreenElements.instance }
 
-      expect(generated_creation_methods).to contain_exactly(:example_page, :example_section)
+    it "creates screen element methods" do
+      generated_creation_methods = screen.public_methods(false)
+
+      expect(generated_creation_methods).to contain_exactly(:example_one, :example_two)
+    end
+
+    it "wraps as expected" do
+      expect(screen.example_one).to be_a_kind_of(Screengem::ScreenElement)
+      expect(screen.example_two).to be_a_kind_of(Screengem::AutomaticVisit)
     end
   end
 end